Renamed and symlinked site script templates
Before this, templates started with numeric prefixes like "05" if
they did not require a specific order of execution. That is nonsense
- "50" makes way more sense!

Also, the preset and stock templates were in different directories. This
created problems if templates have to be moved between directories.

So, now all templates are in the stock directories, and those which
currently happen to be also preset templates are symlinks to the
stock templates of the same name.