Fix path priority ordering
It turned out that it was wrong to suffix the current group with the
next entry.

Instead, if should be prefixed with the new entry.

In order for this to make any sense, the order of group definitions has
to be reversed, too. But the order within the groups stays the same.

However, it turned out that this order was not correct in many entries,
so pretty much everything regarding the order changed.

Change the "+" command into "<---" because now the members of the next
group prefixe rather than suffixe the members of the previous group.