Work around a bug in dash
dash has a problem with HERE-DOCUMENTs on some platforms (not on x86,
but for example on armv7l): When more then one HERE-DOCUMENT in a single
command, command substitution in the HERE-DOCUMENTs won't work.

As a workaround, move the command substitutions out of the HERE-DOCUMENT
into shell variables, and expand the variables in the HERE-DOCUMENT.
Parameter substitution is not buggy.

See also

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=541544
https://www.spinics.net/lists/dash/msg01652.html